requirements.txt
Jinja2
requests
pyyaml
scp
